Create compiler
This is some of the shittest code I have ever written. Is it because I didn't really plan it? Is it because I'm in year 12? Maybe it's because I know nothing about creating a programming language.

I should have designed a simpler language and designed a pipeline for the code.